created an edit route for the users profile, created mandatory address input if logged in with github, sorted form highlighting if fields missing, created delete route for games and users profile, sorted adding comments, deleting them and also allowing the user to reply to specific comments

+++ models/user.js
@@ -23,6 +23,7 @@ userSchema
   .virtual('imageSRC')
   .get(function getImageSRC() {
     if(!this.image) return 'http://euniv.shooliniuniversity.com/erp/required/images/mprofile.png';
+    if(this.image.match(/^http/)) return this.image;
     return `https://s3-eu-west-1.amazonaws.com/wdi-ldn-express-project2/${this.image}`;
   });
 
@@ -44,7 +45,11 @@ userSchema.pre('validate', function checkPassword(next) {
 });
 
 userSchema.pre('remove', function removeImage (next) {
-  s3.deleteObject({ Key: this.image }, next);
+  console.log(this.imageSRC);
+  if(this.imageSRC !== 'http://euniv.shooliniuniversity.com/erp/required/images/mprofile.png') {
+    return s3.deleteObject({ Key: this.image }, next);
+  }
+  next();
 });
 
 userSchema.pre('save', function hashPassword(next) {
